update/refactoring for multiple email support
