+++ server/src/main/resources/password/pwm/config/PwmSetting.xml
@@ -687,19 +687,20 @@
         <regex>^(?!.*all.*)([a-zA-Z][a-zA-Z0-9-]{2,15})$</regex>
         <properties>
             <property key="Minimum">1</property>
+            <property key="Minimum">5</property>
         </properties>
         <default>
-            <value>Email</value>
+            <value>default</value>
         </default>
     </setting>
 
-    <setting hidden="false" key="email.smtp.addresses" level="1">
+    <setting hidden="false" key="email.smtp.address" level="1">
         <regex>^[a-zA-Z0-9.-]*$</regex>
         <default>
             <value />
         </default>
     </setting>
-    <setting hidden="false" key="email.smtp.ports" level="1">
+    <setting hidden="false" key="email.smtp.port" level="1">
         <default>
             <value>25</value>
         </default>
@@ -714,12 +715,12 @@
             <value>noreply@example.org</value>
         </default>
     </setting>
-    <setting hidden="false" key="email.smtp.usernames" level="1">
+    <setting hidden="false" key="email.smtp.username" level="1">
         <default>
             <value />
         </default>
     </setting>
-    <setting hidden="false" key="email.smtp.userpasswords" level="1">
+    <setting hidden="false" key="email.smtp.userpassword" level="1">
     </setting>
 
     <setting hidden="false" key="email.smtp.address" level="1">
@@ -3935,7 +3936,7 @@
     <category hidden="false" key="CHALLENGE_POLICY">
         <profile setting="challenge.profile.list"/>
     </category>
-    <category hidden="false" key="EMAIL_PROFILES">
+    <category hidden="false" key="EMAIL_SERVERS">
         <profile setting="email.profile.list"/>
     </category>
     <category hidden="false" key="EMAIL_PROFILE_SETTING">
