+++ server/src/main/resources/password/pwm/i18n/PwmSetting.properties
@@ -46,9 +46,9 @@ Category_Description_EDIR_CR_SETTINGS=NetIQ eDirectory CR specific settings.
 Category_Description_EDIRECTORY=NetIQ eDirectory specific settings.
 Category_Description_EDIR_SETTINGS=NetIQ eDirectory specific settings.
 Category_Description_EMAIL=<p>Configuration settings for all sent emails.   The settings for the email body configuration are for both plaintext and HTML.  We encourage that for each configured setting and locale for the email body, that you configure both plaintext and HTML. @PwmAppName@ delivers the email in both formats and the email client can choose which to display.</p> <p>Email definitions might use macros.  For more information about macros, see the "View" menu "Show Macro Help".</p>
+Category_Description_EMAIL_SERVERS=Email Servers
 Category_Description_EMAIL_SETTINGS=
 Category_Description_EMAIL_TEMPLATES=
-Category_Description_EMAIL_PROFILES=Email Profile Settings
 Category_Description_FORGOTTEN_USERNAME=Allows a user to search for a forgotten user name using a configurable search filter and attributes.
 Category_Description_GENERAL=General settings for the application.  Settings here control the functionality and behavior of the system overall.
 Category_Description_GUEST=Note\: The guest user registration module requires that the logged in user has sufficient permissions to create users and if so configured, to check for duplicate values.
@@ -144,9 +144,9 @@ Category_Label_EDIR_CR_SETTINGS=eDirectory Challenge Sets
 Category_Label_EDIRECTORY=NetIQ eDirectory
 Category_Label_EDIR_SETTINGS=eDirectory Settings
 Category_Label_EMAIL=Email
+Category_Label_EMAIL_SERVERS=Email Servers
 Category_Label_EMAIL_SETTINGS=Email Settings
 Category_Label_EMAIL_TEMPLATES=Email Templates
-Category_Label_EMAIL_PROFILES=Email Profile Settings
 Category_Label_FORGOTTEN_USERNAME=Forgotten User Name
 Category_Label_GENERAL=Application
 Category_Label_GUEST=Guest Registration
@@ -314,7 +314,7 @@ Setting_Description_email.adminAlert.toAddress=Define this template to send an e
 Setting_Description_email.challenge.token=Define this template to send an email during the forgotten password verification process.  You can use %TOKEN% to insert the token value into the email.
 Setting_Description_email.changePassword=Define this template to send an email to the users when password changes occur. @PwmAppName@ only sends this email when the users change their own passwords.
 Setting_Description_email.changePassword.helpdesk=Define this template to send an email to users when the Help Desk changes the users' passwords. @PwmAppName@ expands macros for this setting based on the user who is changing their password, not the Help Desk user. 
-Setting_Description_email.default.fromAddress=Specify a default From Address for the email templates. This value will be used for all sent email unless indicated in the template.
+Setting_Description_email.default.fromAddress=Specify a default From Address for the email templates.
 Setting_Description_email.deleteAccount=Define this template to send an email to the users after the Account Delete action.
 Setting_Description_email.guest=Define this template to send an email to newly created guest users.
 Setting_Description_email.helpdesk.token=Define this template to send an email during the Help Desk verification process.  You can use %TOKEN% to insert the token value into the email.
@@ -326,11 +326,11 @@ Setting_Description_email.pwExpirationNotice=Email sent to users to notify the u
 Setting_Description_email.queueMaxAge=Specify the maximum age (in seconds) an email can wait in the send queue.  If an email is in the send queue longer than this time, @PwmAppName@ discards it.  Emails only persist in the send queue if there is an IO or network error to the SMTP server while sending the email.
 Setting_Description_email.sendpassword=Define this template to send an email during forgotten password reset process if you enabled the send password functionality.
 Setting_Description_email.sendUsername=Define this template to send an email for the forgotten user name process.
-Setting_Description_email.profile.list=Email profiles for multiple email servers.
-Setting_Description_email.smtp.addresses=Specify an SMTP server address that sends the emails @PwmAppName@ generates.  Removing this setting prevents @PwmAppName@ from sending any emails.  Ensure that the server specified here allows relaying.  For best results, use a local SMTP server.
-Setting_Description_email.smtp.ports=Specify the network port number for the SMTP server.
-Setting_Description_email.smtp.usernames=Specify an SMTP user that logs in to the SMTP server so that it can send the emails @PwmAppName@ generates.  A blank value here sends SMTP messages without authentication.
-Setting_Description_email.smtp.userpasswords=Specify the password for the SMTP user.  A blank value here sends SMTP messages without authentication.
+Setting_Description_email.profile.list=List of SMTP email servers to be used.  @PwmAppName@ will alternate among the servers in the list when a server becomes unreachable.
+Setting_Description_email.smtp.address=Specify an SMTP server address that sends the emails @PwmAppName@ generates.  Removing this setting prevents @PwmAppName@ from sending any emails.  Ensure that the server specified here allows relaying.  For best results, use a local SMTP server.
+Setting_Description_email.smtp.port=Specify the network port number for the SMTP server.
+Setting_Description_email.smtp.username=Specify an SMTP user that logs in to the SMTP server so that it can send the emails @PwmAppName@ generates.  A blank value here sends SMTP messages without authentication.
+Setting_Description_email.smtp.userpassword=Specify the password for the SMTP user.  A blank value here sends SMTP messages without authentication.
 Setting_Description_email.smtp.advancedSettings=Add Name/Value settings to control the behavior of the mail agent. Available settings are defined as part of the <a href\="https\://javamail.java.net/nonav/docs/api/com/sun/mail/smtp/package-summary.html">JavaMail API</a>. The settings must be in "name\=value" format, where name is the key value of a valid JavaMail API setting.
 Setting_Description_email.unlock=Define this template to send an email to users who unlock their own account.
 Setting_Description_email.updateguest=Define this template to send an email to updated guest users.
@@ -813,11 +813,11 @@ Setting_Label_email.pwExpirationNotice=Password Expiration Notification Email
 Setting_Label_email.queueMaxAge=Maximum Email Queue Age
 Setting_Label_email.sendpassword=Send Password Email
 Setting_Label_email.sendUsername=Send User Name Email
-Setting_Label_email.profile.list=Email Profiles
-Setting_Label_email.smtp.addresses=Email Server Address
-Setting_Label_email.smtp.ports=Email Server Port
-Setting_Label_email.smtp.usernames=SMTP Email Server User Name
-Setting_Label_email.smtp.userpasswords=Email Server Password
+Setting_Label_email.profile.list=Email Servers
+Setting_Label_email.smtp.address=SMTP Server Address
+Setting_Label_email.smtp.port=SMTP Server Port
+Setting_Label_email.smtp.username=SMTP Server User Name
+Setting_Label_email.smtp.userpassword=SMTP Server Password
 Setting_Label_email.smtp.advancedSettings=SMTP Email Advanced Settings
 Setting_Label_email.unlock=Unlock Account Email
 Setting_Label_email.updateguest=Guest Registration Update Email
