+++ server/src/test/java/password/pwm/util/queue/EmailQueueManagerTest.java
@@ -22,34 +22,33 @@
 
 package password.pwm.util.queue;
 
-import java.io.IOException;
-import java.util.List;
+import junit.framework.Assert;
+import org.apache.commons.io.IOUtils;
+import org.junit.Test;
+import org.mockito.Mockito;
+import password.pwm.AppProperty;
+import password.pwm.bean.EmailItemBean;
+import password.pwm.config.Configuration;
+import password.pwm.svc.email.EmailServerUtil;
+import password.pwm.svc.email.EmailService;
 
 import javax.mail.Message;
 import javax.mail.MessagingException;
 import javax.mail.internet.InternetAddress;
-
-import junit.framework.Assert;
-
-import org.apache.commons.io.IOUtils;
-import org.junit.Test;
-import org.mockito.Mockito;
-
-import password.pwm.AppProperty;
-import password.pwm.bean.EmailItemBean;
-import password.pwm.config.Configuration;
+import java.io.IOException;
+import java.util.List;
 
 public class EmailQueueManagerTest {
     @Test
     public void testConvertEmailItemToMessage() throws MessagingException, IOException {
-        EmailQueueManager emailQueueManager = new EmailQueueManager();
+        EmailService emailService = new EmailService();
 
         Configuration config = Mockito.mock(Configuration.class);
         Mockito.when(config.readAppProperty(AppProperty.SMTP_SUBJECT_ENCODING_CHARSET)).thenReturn("UTF8");
 
         EmailItemBean emailItemBean = new EmailItemBean("fred@flintstones.tv, barney@flintstones.tv", "bedrock-admin@flintstones.tv", "Test Subject", "bodyPlain", "bodyHtml");
 
-        List<Message> messages = emailQueueManager.convertEmailItemToMessages(emailItemBean, config);
+        List<Message> messages = EmailServerUtil.convertEmailItemToMessages(emailItemBean, config, null);
         Assert.assertEquals(2, messages.size());
 
         Message message = messages.get(0);
